+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This project follows [semantic versioning](http://semver.org).
 
-### UNRELEASED
+### v0.8.2 (2018-01-02)
 
 - [changed] Using serde instead of rustc_serialize
 - [changed] Updated libsodium to 1.0.16